requirements.txt
Django>=4.1
django-debug-toolbar>=3.7.0
django-grappelli>=3.0.3
djangorestframework>=3.14.0
Pillow>=9.3.0
